refactor: directory restructure (#80)
* refactor(styles): moved css files into src/styles directory with consistent naming

* refactor(pages): moved nested pages to pages main directory

* refactor(components): moved nested components into base components directory

* refactor(imports): used vite's path resolution for imports

* refactor: moved pages/Anonymous.jsx to components directory since it's a component and not a page
