+++ client/jsconfig.json
@@ -1,0 +1,29 @@
+{
+    "compilerOptions": {
+        "baseUrl": ".",
+        "paths": {
+            "src/*": [
+                "src/*"
+            ],
+            "components/*": [
+                "src/components/*"
+            ],
+            "pages/*": [
+                "src/pages/*"
+            ],
+            "context/*": [
+                "src/context/*"
+            ],
+            "styles/*": [
+                "src/styles/*"
+            ],
+            "assets/*": [
+                "src/assets/*"
+            ],
+        }
+    },
+    "exclude": [
+        "node_modules",
+        "dist"
+    ]
+}
