+++ client/src/components/NavBar.jsx
