+++ client/src/components/Routes.jsx
@@ -1,10 +1,10 @@
 import { Route, Routes } from 'react-router-dom';
-import ComingSoon from '../../pages/ComingSoon/ComingSoon';
+import ComingSoon from 'pages/ComingSoon';
 
 // Components
 import NavBar from './NavBar';
-import Searching from '../../pages/Searching';
-import Start from '../../pages/Start';
+import Searching from 'pages/Searching';
+import Start from 'pages/Start';
 
 const RoutesComponent = () => {
     return (
