+++ client/src/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from 'react-dom';
 import { BrowserRouter as Router } from 'react-router-dom';
-import './assets/css/index.css'
+import 'styles/index.css'
 import App from './App';
 import { Provider } from 'react-redux';
 import store from './context/redux/store'
